ReOrganize plotting functions, modularize
