+++ fooof/tests/test_plts.py
@@ -3,7 +3,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-from fooof.plts import *
+from fooof.plts.templates import *
 
 ###################################################################################################
 ###################################################################################################
